Create new version of string builder called ahb_sb.
Also introduced a implementation of ahb_sb that offers the possibility
of using arenas to allocate strings. It's incomplete and only the
string_new function has this new feature. Because of this, a new file
has been created for that implementation.
The final version of the library is a single file that can be used with
or without arenas.